+++ grafanalib/core.py
@@ -1175,6 +1175,7 @@ class Graph(Panel):
     :param dataSource: DataSource's name
     :param minSpan: Minimum width for each panel
     :param repeat: Template's name to repeat Graph on
+    :param thresholds: graph panel thresholds
     """
 
     alertThreshold = attr.ib(default=True, validator=instance_of(bool))
@@ -1203,6 +1204,7 @@ class Graph(Panel):
         default=attr.Factory(Tooltip),
         validator=instance_of(Tooltip),
     )
+    thresholds = attr.ib(default=attr.Factory(list))
     xAxis = attr.ib(default=attr.Factory(XAxis), validator=instance_of(XAxis))
     # XXX: This isn't a *good* default, rather it's the default Grafana uses.
     yAxes = attr.ib(
@@ -1238,6 +1240,7 @@ class Graph(Panel):
             'stack': self.stack,
             'steppedLine': self.steppedLine,
             'tooltip': self.tooltip,
+            'thresholds': self.thresholds,
             'type': GRAPH_TYPE,
             'xaxis': self.xAxis,
             'yaxes': self.yAxes,
@@ -2310,7 +2313,8 @@ class StatusmapColor(object):
 
 @attr.s
 class Statusmap(Panel):
-    """Generates json structure for the flant-statusmap-panel visualisation plugin (https://grafana.com/grafana/plugins/flant-statusmap-panel/).
+    """Generates json structure for the flant-statusmap-panel visualisation plugin
+    (https://grafana.com/grafana/plugins/flant-statusmap-panel/).
 
     :param alert
     :param cards: A statusmap card object: keys 'cardRound', 'cardMinWidth', 'cardHSpacing', 'cardVSpacing'
@@ -2402,7 +2406,8 @@ class Svg(Panel):
     :param id: panel id
     :param interval: defines time interval between metric queries
     :param links: additional web links
-    :param reduceCalc: algorithm for reduction to a single value: keys 'mean' 'lastNotNull' 'last' 'first' 'firstNotNull' 'min' 'max' 'sum' 'total'
+    :param reduceCalc: algorithm for reduction to a single value:
+        keys 'mean' 'lastNotNull' 'last' 'first' 'firstNotNull' 'min' 'max' 'sum' 'total'
     :param span: defines the number of spans that will be used for panel
     :param svgFilePath: path to SVG image file to be displayed
     """
@@ -2504,7 +2509,8 @@ class DashboardList(Panel):
     :param transparent: defines if the panel is transparent
 
     :param showHeadings: The chosen list selection (Starred, Recently viewed, Search) is shown as a heading
-    :param showSearch: Display dashboards by search query or tags. Requires you to enter at least one value in Query or Tags
+    :param showSearch: Display dashboards by search query or tags.
+        Requires you to enter at least one value in Query or Tags
     :param showRecent: Display recently viewed dashboards in alphabetical order
     :param showStarred: Display starred dashboards in alphabetical order
     :param maxItems: Sets the maximum number of items to list per section
@@ -2556,7 +2562,8 @@ class Logs(Panel):
     :param showLabels: Show or hide the unique labels column, which shows only non-common labels
     :param showTime: Show or hide the log timestamp column
     :param wrapLogMessages: Toggle line wrapping
-    :param sortOrder: Display results in 'Descending' or 'Ascending' time order. The default is Descending, showing the newest logs first.
+    :param sortOrder: Display results in 'Descending' or 'Ascending' time order. The default is Descending,
+        showing the newest logs first.
     """
     showLabels = attr.ib(default=False, validator=instance_of(bool))
     showTime = attr.ib(default=False, validator=instance_of(bool))
@@ -2585,11 +2592,20 @@ class Logs(Panel):
 
 @attr.s
 class Threshold(object):
-    """Threshold for a gauge
+    """Threshold for for panels
+    (https://grafana.com/docs/grafana/latest/panels/thresholds/)
 
     :param color: color of threshold
     :param index: index of color in gauge
     :param value: when to use this color will be null if index is 0
+    :param op: Graph only - EVAL_LT for less than or EVAL_GT for greater than to indicate what the threshold applies to.
+    :param yaxis: Graph only - Choose left or right for Graph panels
+
+    Example:
+        thresholds = [
+            Threshold('green', 0, 0.0),
+            Threshold('red', 1, 80.0)
+        ]
     """
 
     color = attr.ib()